+++ src/WebJobs.Script/Config/ScriptJobHostOptions.cs
@@ -137,6 +137,11 @@ namespace Microsoft.Azure.WebJobs.Script
         internal TelemetryMode TelemetryMode { get; set; } = TelemetryMode.ApplicationInsights;
 
         /// <summary>
+        /// Gets or sets a value indicating whether the host.json file was created by the host.
+        /// </summary>
+        public bool IsDefaultHostConfig { get; set; }
+
+        /// <summary>
         /// Gets or sets a value indicating the timeout duration for the function metadata provider.
         /// </summary>
         public TimeSpan MetadataProviderTimeout { get; set; } = TimeSpan.Zero;